RoleType move to domain
