+++ src/containers/organization/OrganizationPageContainer.tsx
@@ -2,9 +2,10 @@ import { ApolloError } from '@apollo/client';
 import React, { FC, useMemo } from 'react';
 import { ContributorCardProps } from '../../components/composite/common/cards/ContributorCard/ContributorCard';
 import { isSocialLink, SocialLinkItem } from '../../components/composite/common/SocialLinks/SocialLinks';
+import { RoleType } from '../../domain/user/constants/RoleType';
 import { useOrganization, useUserCardRoleName, useUserContext } from '../../hooks';
 import { useRolesOrganizationQuery } from '../../hooks/generated/graphql';
-import { COUNTRIES_BY_CODE, RoleType } from '../../models/constants';
+import { COUNTRIES_BY_CODE } from '../../models/constants';
 import { CAPABILITIES_TAGSET, KEYWORDS_TAGSET } from '../../models/constants/tagset.constants';
 import { ContainerChildProps } from '../../models/container';
 import { ContributionItem } from '../../models/entities/contribution';
